+++ config.go
@@ -5,10 +5,11 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/portapps/portapps/pkg/utl"
 	"gopkg.in/yaml.v2"
 )
 
-// Configuration holds portapp configuration details
+// Config holds portapp configuration details
 type Config struct {
 	Common Common      `yaml:"common" mapstructure:"common"`
 	App    interface{} `yaml:"app,omitempty" mapstructure:"app"`
@@ -20,9 +21,9 @@ type Common struct {
 }
 
 // loadConfig load common and app configuration
-func loadConfig(appcfg interface{}) error {
-	cfgPath := PathJoin(Papp.Path, fmt.Sprintf("%s.yml", Papp.ID))
-	Papp.config = &Config{
+func (app *App) loadConfig(appcfg interface{}) (err error) {
+	cfgPath := utl.PathJoin(app.RootPath, fmt.Sprintf("%s.yml", app.ID))
+	app.config = &Config{
 		Common: Common{
 			Args: []string{},
 		},
@@ -30,24 +31,25 @@ func loadConfig(appcfg interface{}) error {
 	}
 
 	// Write sample config
-	raw, err := yaml.Marshal(Papp.config)
+	raw, err := yaml.Marshal(app.config)
 	if err != nil {
 		return err
 	}
-	err = ioutil.WriteFile(PathJoin(Papp.Path, fmt.Sprintf("%s.sample.yml", Papp.ID)), raw, 0644)
+	err = ioutil.WriteFile(utl.PathJoin(app.RootPath, fmt.Sprintf("%s.sample.yml", app.ID)), raw, 0644)
 	if err != nil {
 		return err
 	}
 
-	// Check config exists
+	// Skip if config file not found
 	if _, err := os.Stat(cfgPath); err != nil {
 		return nil
 	}
 
-	Log.Info("Loading configuration...")
+	// Read config
 	raw, err = ioutil.ReadFile(cfgPath)
 	if err != nil {
 		return err
 	}
-	return yaml.Unmarshal(raw, &Papp.config)
+
+	return yaml.Unmarshal(raw, &app.config)
 }